script 부분 압축 해체한 후 객체화...

+++ src/main/java/kr/dogfoot/hwplib/reader/docinfo/borderfill/ForFillInfo.java
@@ -129,7 +129,7 @@ public class ForFillInfo {
                 fi.getGradientFill().setBlurringCenter(sr.readUInt1());
             }
         } else {
-            sr.skip(size);
+            sr.skip((int) size);
         }
     }
